requirements.txt
git+https://github.com/eth-brownie/brownie.git@master#egg=eth-brownie
flake8==3.8.4
isort==5.7.0
brownie-token-tester>=0.2.2
